Finally a sane setup for my tab bar.
you have to add the spacing before adding the tabbar, then the 'window'
containing the tab bar is not processed by the spacing layout modifier.
To achieve that, the tab bar has to be defined to the left of the
spacing modifier.

Also added a top bar marking the currently focussed window, inspired
(read: copied) by ethanschoonover's nice xmonad config.